Válaszok átalakítása JSON Objectre, End-pointok átalakítása hiba elkapáshoz és Response-ok beállításához, Service rétegben validálás/ellenőrzés és ez alapján hiba dobása, Tárolt eljárás készítése
getCountyById eljárás létrehozása, hogy ne csak id legyen visszaadva és lehessen validálni ennek segítségével.

+++ Backend/SzakemberKereso/src/main/java/szakemberkereso/Service/AdsCountiesService.java
@@ -5,6 +5,11 @@
 package szakemberkereso.Service;
 
 import java.util.List;
+import java.util.Objects;
+import javax.mail.AuthenticationFailedException;
+import javax.ws.rs.ForbiddenException;
+import szakemberkereso.Configuration.Roles;
+import szakemberkereso.Model.Ads;
 import szakemberkereso.Model.AdsCounties;
 import szakemberkereso.Model.Counties;
 
@@ -14,17 +19,35 @@ import szakemberkereso.Model.Counties;
  */
 public class AdsCountiesService {
         
-    public String addNewCountyToAd(AdsCounties ad_county){
-        String result = AdsCounties.addNewCountyToAd(ad_county);
-        return result;
+    public void addNewCountyToAd(AdsCounties ad_county) throws Exception{
+        //a USER jogosultságú user-ek nem módosíthatnak hirdetéseket
+        if (AuthService.isUserAuthorized(ad_county.getCurrentUserId(), new Roles[]{Roles.ADMIN, Roles.WORKER})){
+            //a user-ek csak a saját hirdetésüket módosíthatják
+            if(!Objects.equals(ad_county.getCurrentUserId(), Ads.getAdsById(ad_county.getAdId()).getUserId())){
+                throw new ForbiddenException("Nincs jogosultsága ehhez a kéréshez.");
+            }
+            AdsCounties.addNewCountyToAd(ad_county);
+        }
+        else{
+            throw new ForbiddenException("Nincs jogosultsága ehhez a kéréshez!");
+        }
     }
     
-    public String deleteCountyFromAd(AdsCounties ad_county){
-        String result = AdsCounties.deleteCountyFromAd(ad_county);
-        return result;
+    public void deleteCountyFromAd(AdsCounties ad_county) throws Exception{
+        //a USER jogosultságú user-ek nem törölhetnek megyét hirdetésből
+        if (AuthService.isUserAuthorized(ad_county.getCurrentUserId(), new Roles[]{Roles.ADMIN, Roles.WORKER})){
+            //a user-ek csak a saját hirdetésüket módosíthatják
+            if(!Objects.equals(ad_county.getCurrentUserId(), Ads.getAdsById(ad_county.getAdId()).getUserId())){
+                throw new ForbiddenException("Nincs jogosultsága ehhez a kéréshez.");
+            }
+            AdsCounties.deleteCountyFromAd(ad_county);
+        }
+        else{
+            throw new ForbiddenException("Nincs jogosultsága ehhez a kéréshez!");
+        }
     }
     
-    public List<Counties> getAllCountiesByAd(Integer ad_id_in){
+    public List<Counties> getAllCountiesByAd(Integer ad_id_in) throws Exception{
         List<Counties> result = AdsCounties.getAllCountiesByAd(ad_id_in);
         return result;
     }
